refactor(mm-next): group profile query and mutation
