+++ packages/mirror-media-next/pages/profile/index.js
@@ -6,7 +6,7 @@ import { useMembership } from '../../context/membership'
 import { useEffect, useState } from 'react'
 import { useRouter } from 'next/router'
 import client from '../../apollo/apollo-client'
-import { fetchMemberProfileByFirebaseId } from '../../apollo/query/profile'
+import { fetchMemberProfileByFirebaseId } from '../../apollo/profile/query/fetch-member-profile'
 import { GCP_PROJECT_ID } from '../../config/index.mjs'
 import errors from '@twreporter/errors'
 import { fetchHeaderDataInDefaultPageLayout } from '../../utils/api'
